+++ analyze_facebook_data.rb
@@ -5,7 +5,10 @@
 require 'nokogiri'
 require 'axlsx'
 
+require_relative 'lib/analyze_facebook_data.rb'
 require_relative 'lib/making_friends_data.rb'
+require_relative 'lib/messages_sent.rb'
+require_relative 'lib/friends_dates.rb'
 
 # images exchanged
 # links exchanged
@@ -28,117 +31,6 @@ require_relative 'lib/making_friends_data.rb'
 # friends gained by month
 # rank everything
 
-class AnalyzeFacebookData
-  attr_accessor :friends, :me, :my_messages_dates, :dictionary, :catalog
-
-  def initialize(data_catalog)
-    @friends           = {}
-    @me                = Hash.new(0)
-    @my_messages_dates = []
-    @dictionary        = Hash.new(0)
-    @catalog           = data_catalog
-  end
-
-  # AnalyzeMessages
-  def start
-    messages_files.each do |file|
-      # open current file
-      content = File.open(file)
-      doc = Nokogiri::HTML(content)
-
-      # get friend name
-      friend_name = doc.title.split('Conversation with ')[1]
-
-      # return value
-      friends[friend_name] ||= {
-        you_count: 0,
-        you_characters: 0,
-        you_words: 0,
-        friend_count: 0,
-        friend_characters: 0,
-        friend_words: 0,
-        total_count: 0
-      }
-
-      # Debug
-      puts "Analyzing conversation with: #{friend_name}"
-
-      # whole conversation
-      conversation = doc.css('.thread').children
-
-      # instance
-      current_message_sender = ''
-
-      conversation.each do |conversation_node|
-        # is the converstation a message?
-        if conversation_node.name == 'div' && conversation_node['class'] == 'message'
-          # user sending this message
-          current_message_sender  = conversation_node.css('span.user').text
-          # meta conversation data
-          date_info               = conversation_node.css('span.meta').text
-
-          # who sent message, was this me or you
-          if current_message_sender == user_name
-            # me
-            me[:total_message_count]            += 1
-            friends[friend_name][:you_count]    += 1
-            friends[friend_name][:total_count]  += 1
-            my_messages_dates << Time.parse(date_info)
-          else
-            # a friend of mine
-            friends[friend_name][:total_count] += 1
-            friends[friend_name][:friend_count] += 1
-          end
-        end
-
-        # skip this conversation node
-        next unless conversation_node.name == 'p'
-
-        # numbers about words in the conversation
-        paragraph        = conversation_node.text.downcase
-        paragraph        = paragraph.delete(',').delete('.')
-        paragraph_length = paragraph.length
-        paragraph_words  = paragraph.split(' ')
-
-        # who sent this message, me or you?
-        if current_message_sender == user_name
-          # me
-          me[:total_characters]   += paragraph_length
-          me[:total_words]        += paragraph_words.length
-          me[:total_xd]           += paragraph.scan('xd').length
-
-          # build dictionary
-          paragraph_words.each do |word|
-            dictionary[word] += 1
-          end
-
-          # conversation with friend
-          friends[friend_name][:you_characters] += paragraph_length
-          friends[friend_name][:you_words]      += paragraph_words.length
-        else
-          # you
-          friends[friend_name][:friend_characters]  += paragraph_length
-          friends[friend_name][:friend_words]       += paragraph_words.length
-        end
-      end
-    end
-    self
-  end
-
-  def ranking
-    friends.sort_by { |_name, friend| friend[:total_count] }.reverse
-  end
-
-  def messages_files
-    Dir.glob("#{catalog}/messages/*.html")[0..5]
-  end
-
-  def user_name
-    Nokogiri::HTML(File.open("#{catalog}/index.htm")).title
-                                                     .split(' - Profile')[0]
-  end
-end
-
 analyze_facebook_data = AnalyzeFacebookData.new(ARGV[0]).start
 
 # CreatePackage
@@ -154,44 +46,6 @@ package.workbook.add_worksheet(name: 'Friends ranking') do |sheet|
                    friend_data[:friend_characters], friend_data[:you_words],
                    friend_data[:friend_words]]
     rank += 1
-  end
-end
-
-# analyze message patterns when messages are sent
-class MessagesSent
-  attr_accessor :by_month, :by_year, :by_day_of_week,
-                :by_weekend, :by_date, :by_hour,
-                :by_year_hour
-  attr_reader :my_messages_dates
-
-  def initialize(my_messages_dates)
-    @my_messages_dates = my_messages_dates
-    @by_month       = Hash.new(0)
-    @by_year        = Hash.new(0)
-    @by_day_of_week = Hash.new(0)
-    @by_weekend     = Hash.new(0)
-    @by_date        = Hash.new(0)
-    @by_hour        = Hash.new(0)
-    @by_year_hour   = Hash.new(0)
-  end
-
-  def build
-    my_messages_dates.each do |date|
-      by_month[date.strftime('%B')] += 1
-      by_year[date.year] += 1
-      by_day_of_week[date.strftime('%A')] += 1
-
-      if date.friday? || date.saturday? || date.sunday?
-        by_weekend[:weekend] += 1
-      else
-        by_weekend[:working] += 1
-      end
-
-      by_date[date.strftime('%F')] += 1
-      by_hour[date.hour] += 1
-      by_year_hour["#{date.year} - #{date.hour}"] += 1
-    end
-    self
   end
 end
 
@@ -319,62 +173,6 @@ Dir.chdir("#{analyze_facebook_data.catalog}/html/") do
     contacts.sort_by { |contact_name, _info| contact_name }.each do |contact_name, contact_num|
       sheet.add_row [contact_name, contact_num]
     end
-  end
-end
-
-class FriendsDates
-  # analyze making of friends
-  # Returns an array of [<Date>,<Date>]
-
-  attr_reader :file
-  attr_accessor :friends_dates
-
-  def self.analyze(catalog)
-    new(catalog).analyze
-  end
-
-  def initialize(catalog)
-    @friends_dates = []
-    @file = "#{catalog}/html/friends.htm"
-  end
-
-  def analyze
-    friends_list.each do |friend_element|
-      if friend_with_email(friend_element)
-        _name, date_added = friend_with_email(friend_element).captures
-      else
-        _name, date_added = friend_element.text
-                                          .match(/(.*)\s\((.*)\)/)
-                                          .captures
-      end
-
-      date = if date_added == 'Today'
-               Date.today
-             elsif date_added == 'Yesterday'
-               Date.today.prev_day
-             else
-               Date.parse(date_added)
-             end
-
-      friends_dates << date
-    end
-    self
-  end
-
-  def content
-    File.open(file).read
-  end
-
-  def doc
-    Nokogiri::HTML(content)
-  end
-
-  def friends_list
-    doc.css('div.contents > ul')[0].css('li')
-  end
-
-  def friend_with_email(friend_element)
-    friend_element.text.match(/(.*)\s\((.*)\)\s\((.*)\)/)
   end
 end
